attach and detach run. Run refactoring
