+++ rust/crates/sift_stream/src/stream/mode/ingestion_config.rs
@@ -1,7 +1,10 @@
 use super::super::{
     RetryPolicy, SiftStream, SiftStreamMode, channel::ChannelValue, time::TimeValue,
 };
-use crate::backup::{BackupsManager, DiskBackupsManager, InMemoryBackupsManager};
+use crate::{
+    backup::{BackupsManager, DiskBackupsManager, InMemoryBackupsManager},
+    stream::run::{RunSelector, load_run_by_form, load_run_by_id},
+};
 use futures_core::Stream;
 use prost::Message;
 use sift_connect::SiftChannel;
@@ -728,6 +731,27 @@ impl SiftStream<IngestionConfigMode> {
             );
         }
         Ok(())
+    }
+
+    /// Attach a run to the stream. Any data provided through [SiftStream::send] after return
+    /// of this function will be associated with the run.
+    pub async fn attach_run(&mut self, run_selector: RunSelector) -> Result<()> {
+        let run = match run_selector {
+            RunSelector::ById(run_id) => load_run_by_id(self.grpc_channel.clone(), &run_id).await?,
+            RunSelector::ByForm(run_form) => {
+                load_run_by_form(self.grpc_channel.clone(), run_form).await?
+            }
+        };
+
+        self.mode.run = Some(run);
+
+        Ok(())
+    }
+
+    /// Detach the run, if any, associated with the stream. Any data provided through [SiftStream::send] after
+    /// this function is called will not be associated with a run.
+    pub fn detach_run(&mut self) {
+        self.mode.run = None;
     }
 
     /// This will conclude the stream and return when Sift has sent its final response. It is
